split the dimension manipulation code into the transforms module
