+++ aicsimageio/tests/testAICSImage.py
@@ -1,15 +1,10 @@
-from pathlib import Path
 import random
+from pathlib import Path
 
 import numpy as np
 import pytest
 
 from aicsimageio.aics_image import AICSImage
-from aicsimageio.exceptions import ConflictingArgsError
-
-
-TEST_DATA_DIR = Path(__file__).parent / "img"
-TEST_OME = TEST_DATA_DIR / "img40_1.ome.tif"
 
 
 class ImgContainer(object):
@@ -51,34 +46,6 @@ def test_helper_class(example_img5):
     assert example_img5.remap("XYZCT") == [4, 3, 2, 1, 0]
 
 
-# test _reshape_data which is called when data is accessed
-@pytest.mark.parametrize("kwargs,expected", [
-    ({'data': np.zeros((10, 1, 5, 6, 200, 400)), 'given_dims': "STCZYX", 'return_dims': "CSZYX"}, (5, 10, 6, 200, 400)),
-    ({'data': np.zeros((6, 200, 400)), 'given_dims': "ZYX", 'return_dims': "STCZYX"}, (1, 1, 1, 6, 200, 400)),
-    ({'data': np.zeros((6, 200, 400)), 'given_dims': "ZYX", 'return_dims': "ZCYSXT"}, (6, 1, 200, 1, 400, 1)),
-    pytest.param({'data': np.zeros((6, 200, 400)), 'given_dims': 'ZYX', 'return_dims': 'TYXC', 'Z': 7}, 5,
-                 marks=pytest.mark.raises(exception=IndexError)),
-    pytest.param({'data': np.zeros((6, 200, 400)), 'given_dims': 'ZYX', 'return_dims': 'TYXCZ', 'Z': 7}, 5,
-                 marks=pytest.mark.raises(exception=ConflictingArgsError)),
-])
-def test_reshape_data(kwargs, expected):
-    ans = AICSImage._reshape_data(**kwargs)
-    assert ans.shape == expected
-
-
-@pytest.mark.parametrize("kwargs,expected", [
-    ({'data': np.zeros((1, 2, 3, 4, 5, 6)), 'given_dims': "STCZYX", 'return_dims': "XYZCTS"}, (6, 5, 4, 3, 2, 1)),
-    ({'data': np.zeros((1, 2, 3)), 'given_dims': "ZYX", 'return_dims': "ZXY"}, (1, 3, 2)),
-    pytest.param({'data': np.zeros((6, 200, 400)), 'given_dims': 'ZYX', 'return_dims': 'TYXC'}, 5,
-                 marks=pytest.mark.raises(exception=ConflictingArgsError)),
-    pytest.param({'data': np.zeros((6, 200, 400)), 'given_dims': 'ZYX', 'return_dims': 'TYXCZ'}, 5,
-                 marks=pytest.mark.raises(exception=ConflictingArgsError)),
-])
-def test_transpose_to_dims(kwargs, expected):
-    data = AICSImage._transpose_to_dims(**kwargs)
-    assert data.shape == expected
-
-
 @pytest.mark.skip(reason='Waiting on ndarray support in DefaultReader')
 def test_transposed_output(example_img5):
     output_array = example_img5.image.get_image_data("XYZCT")
@@ -112,14 +79,15 @@ def test_get_image_data_small_data(example_img3ctx):
 
 
 @pytest.mark.parametrize("filepath", [
-        TEST_OME,
-        str(TEST_OME),
-        pytest.param(TEST_DATA_DIR, marks=pytest.mark.raises(exception=IsADirectoryError)),
+        "img40_1.ome.tif",
+        str("img40_1.ome.tif"),
+        pytest.param(None, marks=pytest.mark.raises(exception=IsADirectoryError)),
         pytest.param("fakeimage.ome.tif", marks=pytest.mark.raises(exception=FileNotFoundError)),
         # pytest.param(b"not-a-valid-image-byte-array", marks=pytest.mark.raises(exception=ValueError)),
-        pytest.param("/This/is/a/bogus/file.ome.tif", marks=pytest.mark.raises(exception=FileNotFoundError)),
+        pytest.param("a/bogus/file.ome.tif", marks=pytest.mark.raises(exception=FileNotFoundError)),
 ])
-def test_file_exceptions(filepath):
+def test_file_exceptions(image_dir, filepath):
+    filepath = image_dir if filepath is None else filepath
     image = AICSImage(filepath)
     assert image is not None
