+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+markers =
+    raises: this registers pytest.mark.raises thus preventing related warnings
